Make a major fix to how tokens expire
MongoDB only lets you use its built in TTL feature if the field it is comparing against is a Date type.
The Sails ORM does not include support for a datatype, so the main change in this commit is changing the "expires" attribute to a "ref" type and adding a check to make sure it is a Date object.