Add fillAttributes method in case of missing attributes in production
In production, Sails will not alter database records to have new attributes added with updates to the software.
This often results in undefined attributes that cause issues with rendering models in views. I can write migration scripts to fix this later, but as a failsafe, fillAttributes replaces undefined attributes with their default values as defined in the models. I'm a little annoyed that Sails can't just do this for me, but at least the problem is fixed.